services/blog-worker/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.2.0
langchain-core>=0.3.0
xgboost>=2.0.0
lightgbm>=4.3.0
scikit-learn>=1.4.0
numpy>=1.24.0
pandas>=2.1.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
openai>=1.0.0
supabase>=2.4.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
mlflow>=2.16.0
click>=8.1.0
